refactor: restructure code
fix #12
