+++ src/controller/imageByExtensionDimensionColor.js
@@ -1,0 +1,15 @@
+import mime from 'mime-types'
+
+import { createImage } from '../model/createImage.js'
+import { createImageOptions, extractRequestParametersToCreateImageOptions } from '../model/imageGeneration.js'
+
+export const imageByExtensionDimensionColor = (req, res) => {
+  extractRequestParametersToCreateImageOptions(req.params)
+
+  createImageOptions.callbackOnGeneratedResponseObject = res
+
+  res
+    .status(200)
+    .contentType(mime.lookup(req.params.extension))
+  createImage(createImageOptions)
+}
